[SPARK-46094][CONNECT] Add support for code profiling executors
This adds support for the async profiler to Spark

Profiling of JVM applications on a cluster is cumbersome and it can be complicated to save the output of the profiler especially if the cluster is on K8s where the executor pods are removed and any files saved to the local file system become inaccessible.
This feature makes it simple to turn profiling on/off, includes the jar/binaries needed for profiling,  and makes it simple to save output to an HDFS location.

This PR introduces three new configuration parameters. These are described in the documentation.