[SPARK-44945][DOCS][PYTHON] Automate PySpark error class documentation
### What changes were proposed in this pull request?

This PR introduces a new script named `errors_doc_gen.py` that automatically generates an RST documentation file containing the error classes defined in `pyspark.errors.error_classes`.

### Why are the changes needed?

With the increasing number of error classes in PySpark, it becomes imperative to have a centralized documentation that provides an overview of these errors for both developers and users. This automated script ensures that the documentation is always up-to-date with the error classes defined in the source code, reducing the manual effort required to update the documentation each time a new error class is introduced or an existing one is modified.

### Does this PR introduce _any_ user-facing change?

No, this script is primarily for internal use by developers to generate documentation. End-users will benefit from the updated documentation that this script produces, but they will not directly interact with the script itself.

### How was this patch tested?

The script was executed locally, and the generated RST file was verified for accuracy and completeness. Furthermore, the generated RST was rendered using Sphinx to ensure it produces the expected visual output.

<img width="1395" alt="Screenshot 2023-08-24 at 10 19 52 PM" src="https://github.com/apache/spark/assets/44108233/00142374-2fba-43c1-9034-c5db6989cffa">

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #42658 from itholic/err_doc_automation.

Authored-by: itholic <haejoon.lee@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>