[SPARK-44860][SQL] Add SESSION_USER function
### What changes were proposed in this pull request?

This change implements `SESSION_USER` expression. It behaves exactly the same as `CURRENT_USER` but according to standard when respective function is used inside a routine (UDF):
- `CURRENT_USER` should return security definer, i.e. owner of an UDF
- `SESSION_USER` should return connected user.

The code is duplicated for this reason - to be able to identify which expression is used inside a routing.

### Why are the changes needed?

This is a missing expression defined by SQL standard.

### Does this PR introduce _any_ user-facing change?

Yes, this change introduces a new expression.

### How was this patch tested?

Updating existing unit tests.

Closes #42549 from vitaliili-db/session_user.

Authored-by: Vitalii Li <vitalii.li@databricks.com>
Signed-off-by: Kent Yao <yao@apache.org>