[SPARK-44971][PYTHON] StreamingQueryProgress event fromJson bug fix
### What changes were proposed in this pull request?

The `fromJson` method for `StreamingQueryProgress` excepts the field `batchDuration` is in the dict.
That method is used internally for converting a json representation of `StreamingQueryProgress` into python object, commonly created in the Scala side `json` method of the same object.

But the `batchDuration` field is not there before https://github.com/apache/spark/pull/42077, which is only merged to 4.0. Therefore we add a catch there to prevent this method from failing.

### Why are the changes needed?

Necessary bug fix

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Added unit test

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #42686 from WweiL/SPARK-44971-fromJson-bugfix.

Lead-authored-by: Wei Liu <wei.liu@databricks.com>
Co-authored-by: Wei Liu <z920631580@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>