[SPARK-45012][SQL] CheckAnalysis should throw inlined plan in AnalysisException
### What changes were proposed in this pull request?

CheckAnalysis should throw inlined plan in AnalysisException

### Why are the changes needed?

Before this change, the plan attached to AnalysisException is analyzed plan but not inlined. However, `CheckAnalysis` inlines and checks the plan, so if an exception is from `CheckAnalysis`, it should attach the inlined version of plan which is more useful to debug than the original analyzed plan, especially when there is CTE in inside the plan.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing UT

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #42729 from amaliujia/improve_analyzer.

Authored-by: Rui Wang <rui.wang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>