Revert "[SPARK-43646][CONNECT][TESTS] Make both SBT and Maven use `spark-proto` uber jar to test the `connect` module"
### What changes were proposed in this pull request?
This reverts commit df63adf734370f5c2d71a348f9d36658718b302c.

### Why are the changes needed?
As [reported](https://github.com/apache/spark/pull/42236#issuecomment-1700493815) by MaxGekk , the solution for https://github.com/apache/spark/pull/42236 is not perfect, and it breaks the usability of importing Spark as a Maven project into idea. On the other hand, if `mvn clean test` is executed, test failures will also occur like
```
[ERROR] [Error] /tmp/spark-3.5.0/connector/connect/server/target/generated-test-sources/protobuf/java/org/apache/spark/sql/protobuf/protos/TestProto.java:9:46:  error: package org.sparkproject.spark_protobuf.protobuf does not exist
```
Therefore, this pr will revert the change of SPARK-43646, and `from_protobuf messageClassName` and `from_protobuf messageClassName options` in `PlanGenerationTestSuite` will be ignored in a follow-up.

At present, it is difficult to make the maven testing of the `spark-protobuf` function in the `connect` module as good as possible.

### Does this PR introduce _any_ user-facing change?
NO

### How was this patch tested?
Pass GitHub Actions

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #42746 from LuciferYang/Revert-SPARK-43646.

Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: yangjie01 <yangjie01@baidu.com>