[SPARK-45020][SQL]: Supress the org.apache.spark.sql.catalyst.analysis.NoSuchDatabaseException: Database 'default' not found (state=08S01,code=0) as the 'default' database always exists in the system and cannot be dropped by anyone.
So in the method requireDbExist(db), currently there is no check to ensure if its for
non-default db. And thus when a spark 3.1 cluster is created using shared metastore with spark 2.4
cluster, its throwing an alert saying that the default db doesn't exist. Thus we need to
suppress this exception.