[SPARK-44833][CONNECT] Fix sending Reattach too fast after Execute
### What changes were proposed in this pull request?

Redo the retry logic, so that getting a new iterator via ReattachExecute does not depend on "firstTry", but there is logic in "callIter" with unsetting the iterator when a new one is needed.

### Why are the changes needed?

After an "INVALID_HANDLE.OPERATION_NOT_FOUND" error, client would realize that the failure in ReattachExecute was because the initial ExecutePlan didn't reach the server. It would then call another ExecutePlan, and it will throw a RetryException to let the retry logic handle retrying. However, the retry logic would then immediately send a ReattachExecute, and the client will want to use the iterator of the reattach.

However, on the server the ExecutePlan and ReattachExecute could race with each other:
* ExecutePlan didn't reach executeHolder.runGrpcResponseSender(responseSender) in SparkConnectExecutePlanHandler yet.
* ReattachExecute races around and reaches executeHolder.runGrpcResponseSender(responseSender) in SparkConnectReattachExecuteHandler first.
* When ExecutePlan reaches executeHolder.runGrpcResponseSender(responseSender), and executionObserver.attachConsumer(this) is called in ExecuteGrpcResponseSender of ExecutePlan, it will kick out the ExecuteGrpcResponseSender of ReattachExecute.

So even though ReattachExecute came later, it will get interrupted by the earlier ExecutePlan and finish with a INVALID_CURSOR.DISCONNECTED error.

After this change, such a race between ExecutePlan / ReattachExecute can still happens, but the client should no longer send these requests in such quick succession.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Integration testing.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #42806 from juliuszsompolski/SPARK-44833.

Authored-by: Juliusz Sompolski <julek@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>