[MINOR][DOCS] Change "filter" to "transform" in `transform` function docstring
### What changes were proposed in this pull request?
This PR proposes a simple change in the documentation for the `transform` function in `sql`. I believe where it currently reads "filter" it should read "transform".

### Why are the changes needed?
I believe this change might not be needed per se, but it would be a slight improvement to the current version to avoid the misnomer.

### Does this PR introduce _any_ user-facing change?
Yes, it shows the word "transform' instead of "filter" in the documentation for the `transform` `sql`. function.

### How was this patch tested?
This patch was not tested because it only changes documentation.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #42858 from gdahia/patch-1.

Lead-authored-by: Gabriel Dahia <gdahia@protonmail.com>
Co-authored-by: Gabriel Dahia <gdahia@users.noreply.github.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>