[SPARK-45069][SQL] SQL variable should always be resolved after outer reference
### What changes were proposed in this pull request?

This is a bug fix for the recently added SQL variable feature. It's designed to resolve columns to SQL variable as the last resort, but for columns in Aggregate, we may resolve columns to outer reference first.

### Why are the changes needed?

bug fix

### Does this PR introduce _any_ user-facing change?

yes, the query result can be wrong before this fix

### How was this patch tested?

new tests

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #42803 from cloud-fan/meta-col.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>