[SPARK-45157][SQL] Avoid repeated `if` checks in `[On|Off|HeapColumnVector`
### What changes were proposed in this pull request?

This is a small followup of https://github.com/apache/spark/pull/42850. `getBytes` checks if the `dictionary` is null or not, then call `getByte` which also checks if the `dictionary` is null or not. This PR avoids the repeated if checks by copying one line code from `getByte` to `getBytes`. The same applies to other `getXXX` methods.

### Why are the changes needed?

Make the perf-critical path more efficient.

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

existing tests

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #42903 from cloud-fan/vector.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>