[SPARK-45084][SS] StateOperatorProgress to use accurate effective shuffle partition number
### What changes were proposed in this pull request?
Make StateOperatorProgress.numShufflePartitions to use the effective number of shuffle partitions is reported.
This metric StateStoreWriter.numShufflePartitions is dropped at the same time, as it is not a metric anymore.

### Why are the changes needed?
Currently, there is a numShufflePartitions "metric" reported in
StateOperatorProgress part of the progress report. However, the number is reported by aggregating executors so in the case of task retry or speculative executor, the metric is higher than number of shuffle partitions for the query plan. We change the metric to use the value to use to make it more usable.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
StreamingAggregationSuite contains a unit test that validates the value

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #42822 from siying/numShufflePartitionsMetric.

Authored-by: Siying Dong <siying.dong@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>