[SPARK-45146][DOCS] Update the default value of 'spark.executor.logs.rolling.strategy'
What changes were proposed in this pull request?
The PR updates the default value of 'spark.executor.logs.rolling.strategy in configuration.html on the website

Why are the changes needed?
The default value of 'spark.executor.logs.rolling.strategy' is '', but is not (none). The website has Inaccurate description.It can bring ambiguity.

Does this PR introduce any user-facing change?
No

How was this patch tested?
It doesn't need to.

Was this patch authored or co-authored using generative AI tooling?
No

Closes #42919 from chenyu-opensource/branch-SPARK-45160.

Authored-by: chenyu <119398199+chenyu-opensource@users.noreply.github.com>
Signed-off-by: Sean Owen <srowen@gmail.com>