[SPARK-45138][SS] Define a new error class and apply it when checkpointing state to DFS fails
### What changes were proposed in this pull request?

In this change, we add a new a new error class when checkpointing state to the DFS, for either state store provider, fails during `commit`.

### Why are the changes needed?

Users might be confused when they see an `IOException`, for example, if a call to `commit` fails. This is a neat and self-explanatory wrapper around such exceptions.

### Does this PR introduce _any_ user-facing change?

Yes, the error message when DFS checkpoint fails is now a `SparkException` with error class.

### How was this patch tested?

- Modified existing UTs

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #42895 from neilramaswamy/nr-state-files-commit-logging.

Authored-by: Neil Ramaswamy <neil.ramaswamy@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>