[MINOR][SQL][DOCS] Remove JDK 8 related information in the comemnts for aes_encrypt and aes_decrypt
### What changes were proposed in this pull request?

This PR proposes to fix the comments in both `aes_encrypt` and `aes_decrypt`. Did a quick check for Scala/Python/R API, and seems like this is the only place to fix.

### Why are the changes needed?

We dropped JDK 8 at SPARK-44112

### Does this PR introduce _any_ user-facing change?

No, it's Scaladoc, and the doc is not user-facing.

### How was this patch tested?

Build in this CI should check them.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43091 from HyukjinKwon/minor-doc-AesEncrypt.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>