[SPARK-45308][PS][TESTS] Enable `GroupbySplitApplyTests.test_split_apply_combine_on_series` for pandas 2.0.0
### What changes were proposed in this pull request?

This PR proposes to enable `GroupbySplitApplyTests.test_split_apply_combine_on_series`.

### Why are the changes needed?

Similar to https://github.com/apache/spark/pull/43002, this test is skipped since Pandas 2.0.0 upgrade, but the root cause of the test failure is classified as regression from Pandas. So we can manually make the test pass for now and will update the test when Pandas regression is resolved.

### Does this PR introduce _any_ user-facing change?

No, it's test-only.

### How was this patch tested?

To update the existing test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43096 from itholic/SPARK-45308.

Authored-by: Haejoon Lee <haejoon.lee@databricks.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>