[SPARK-45376][CORE] Add netty-tcnative-boringssl-static dependency
### What changes were proposed in this pull request?

This PR adds a new dependency on this library so that we can use it for SSL functionality.

We also include the network-common test helper library in a few places, which will be needed for reusing some test configuration files in the future.

### Why are the changes needed?

These dependency changes are needed so we can use this library to provide SSL functionality, and for test helpers.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

CI. This is used in unit tests and functionality as part of https://github.com/apache/spark/pull/42685, from which this is being split out

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #43164 from hasnain-db/spark-tls-deps.

Authored-by: Hasnain Lakhani <hasnain.lakhani@databricks.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>