[SPARK-45425][SQL] Mapped TINYINT to ShortType for MsSqlServerDialect
### What changes were proposed in this pull request?
Mapped TINYINT to `ShortType` for MsSqlServerDialect. For TINYINT of sql server, its range is `0-255`

### Why are the changes needed?
TINYINT of SQL server is not correctly mapped to spark types when using JDBC connector. For now, it is mapped to the `IntegerType`, which is not accurate. It should be mapped to `ShortType`.

### Does this PR introduce _any_ user-facing change?
Yes, after the change, `TINYINT` of SQL server will be mapped to `ShortType`

### How was this patch tested?
UT

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #43232 from zeruibao/SPARK-45425-map-tinyint.

Authored-by: zeruibao <zerui.bao@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>