[SPARK-45457][DOCS] Surface sc.setLocalProperty() value=NULL param meaning
### What changes were proposed in this pull request?

sc.setLocalProperty() has special meaning/feature for `null` value as a parameter and when supplied then removes the property associated with the key parameter. It's only mentioned in the [Fair Scheduler section of the documentation](https://spark.apache.org/docs/latest/job-scheduling.html#fair-scheduler-pools) although it is not limited to scheduler pool config and has got more general use cases.

So, it would be nice to document that on the APIs as well for users' benefit.

### Why are the changes needed?

API documentation

### Does this PR introduce _any_ user-facing change?
Yes

### How was this patch tested?
Local lint

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #43269 from khalidmammadov/doc_set_local_prop.

Authored-by: khalidmammadov <khalidmammadov9@gmail.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>