[SPARK-45401][PYTHON] Add a new method `cleanup` in the UDTF interface
### What changes were proposed in this pull request?

This PR adds a new API `cleanup` to the current Python UDTF interface.

### Why are the changes needed?

Currently, the `terminate` method of a UDTF is always executed, regardless of whether the `eval` method calls are successful. This is problematic as users might have certain logic in `terminate` that can only be executed once all `eval` calls succeed.

But what if users wish to perform cleanup actions during UDTF execution, such as closing connections? One option is for users to embed a try...except logic within the eval call:

```
def eval(self, row: Any):
  try:
    run_code()
  except Exception:
    clean_up()
```

However, running this try-except block for every eval call can be expensive, potentially affecting the performance of UDTFs.

To tackle this, we can introduce a new method in the UDTF interface that will be called regardless of the outcome. Now the logic would look like:

```
try:
  for row in rows:
    eval(row)
  terminate()
finally:
  cleanup()
```

### Does this PR introduce _any_ user-facing change?

Yes.

### How was this patch tested?

Unit test

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #43225 from allisonwang-db/spark-45401-udtf-cleanup.

Authored-by: allisonwang-db <allison.wang@databricks.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>