[SPARK-45310][CORE] Report shuffle block status should respect shuffle service during decommission migration
### What changes were proposed in this pull request?
Currently, when migrating a shuffle block during decommission, its location type would be changed from shuffle service (if enabled) to the executor. This is of course not the expected behavior. We should still use the external shuffle service to fetch the migrated shuffle blocks. This PR fixes this issue by respecting the shuffle service when reporting the shuffle blocks.

### Why are the changes needed?

Bug fix/improvment.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?

Unit test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43112 from Ngone51/shuffle-migration-report.

Authored-by: Yi Wu <yi.wu@databricks.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>