[SPARK-44855][CONNECT] Small tweaks to attaching ExecuteGrpcResponseSender to ExecuteResponseObserver
### What changes were proposed in this pull request?

Small improvements can be made to the way new ExecuteGrpcResponseSender is attached to observer.
* Since now we have addGrpcResponseSender in ExecuteHolder, it should be ExecuteHolder responsibility to interrupt the old sender and that there is only one at a time, and to ExecuteResponseObserver's responsibility
* executeObserver is used as a lock for synchronization. An explicit lock object could be better.

Fix a small bug, when ExecuteGrpcResponseSender will not be waken up by interrupt if it was sleeping on the grpcCallObserverReadySignal. This would result in the sender potentially sleeping until the deadline (2 minutes) and only then removed, which would potentially delay timing the execution out by these 2 minutes. It should **not** cause any hang or wait on the client side, because if ExecuteGrpcResponseSender is interrupted, it means that the client has already came back with a new reattach, and the old sender is being kicked out.

### Why are the changes needed?

Minor cleanup of previous work.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing tests in ReattachableExecuteSuite.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43181 from juliuszsompolski/SPARK-44855.

Authored-by: Juliusz Sompolski <julek@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>