[SPARK-45402][SQL][PYTHON] Add UDTF API for 'eval' and 'terminate' methods to consume previous 'analyze' result
### What changes were proposed in this pull request?

This PR adds a Python UDTF API for the `eval` and `terminate` methods to consume the previous `analyze` result.

This also works for subclasses of the `AnalyzeResult` class, allowing the UDTF to return custom state from `analyze` to be consumed later.

For example, we can now define a UDTF that perform complex initialization in the `analyze` method and then returns the result of that in the `terminate` method:

```
def MyUDTF(self):
    dataclass
    class AnalyzeResultWithBuffer(AnalyzeResult):
        buffer: str

    udtf
    class TestUDTF:
        def __init__(self, analyze_result):
            self._total = 0
	    self._buffer = do_complex_initialization(analyze_result.buffer)

        staticmethod
        def analyze(argument, _):
            return AnalyzeResultWithBuffer(
                schema=StructType()
		    .add("total", IntegerType())
		    .add("buffer", StringType()),
                with_single_partition=True,
                buffer=argument.value,
            )

        def eval(self, argument, row: Row):
            self._total += 1

        def terminate(self):
            yield self._total, self._buffer

self.spark.udtf.register("my_ddtf", MyUDTF)
```

Then the results might look like:

```
sql(
    """
    WITH t AS (
      SELECT id FROM range(1, 21)
    )
    SELECT total, buffer
    FROM test_udtf("abc", TABLE(t))
    """
).collect()

> 20, "complex_initialization_result"
```

### Why are the changes needed?

In this way, the UDTF can perform potentially expensive initialization logic in the `analyze` method just once and result the result of such initialization rather than repeating the initialization in `eval`.

### Does this PR introduce _any_ user-facing change?

Yes, see above.

### How was this patch tested?

This PR adds new unit test coverage.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43204 from dtenedor/prepare-string.

Authored-by: Daniel Tenedorio <daniel.tenedorio@databricks.com>
Signed-off-by: Takuya UESHIN <ueshin@databricks.com>